% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djt.R
\name{baixar_djt}
\alias{baixar_djt}
\title{Baixar links do djt.}
\usage{
baixar_djt(regiao = 0L, caderno = "", dt_min = Sys.Date() - 1,
  dt_max = dt_min, dir = "data-raw/djt")
}
\arguments{
\item{regiao}{inteiro entre 1 e 24 identificando o TRT.}

\item{caderno}{\code{jud} para judicial, \code{adm} para administrativo. Default \code{jud}.}

\item{dt_min, dt_max}{datas minima e maxima da busca, em formato \code{Date} ou 'YYYY-MM-DD'. Default \code{\link{Sys.Date}}}

\item{dir}{pasta onde os arquivos serao salvos.}
}
\value{
\code{tibble} com identificadores e links dos DJTs.
}
\description{
Baixar os pdfs do DJT com base em datas, caderno e regiao.
}

