% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{chop_at}
\alias{chop_at}
\title{Chop string at given positions}
\usage{
chop_at(string, positions)
}
\arguments{
\item{string}{Input string}

\item{positions}{Object returned by \code{\link[=find_all]{find_all()}}}
}
\description{
Chop string at given positions
}
\seealso{
\code{\link[=find_all]{find_all()}}
}
