% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special-cleaners.R
\name{clean_all_but_first}
\alias{clean_all_but_first}
\title{Special cleaner
trim all texts but keeps left trim on first element}
\usage{
clean_all_but_first(x)
}
\description{
Special cleaner
trim all texts but keeps left trim on first element
}
