% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special-cleaners.R
\name{clean_right_side}
\alias{clean_right_side}
\title{Special cleaner
trim just the right side of a text and remove numbers in the end.
Doesn't clean if num is NA}
\usage{
clean_right_side(x, na_num)
}
\description{
Special cleaner
trim just the right side of a text and remove numbers in the end.
Doesn't clean if num is NA
}
