% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-section.R
\name{djt_parse}
\alias{djt_parse}
\title{Parse DJT files}
\usage{
djt_parse(pdf_file, verbose = TRUE)
}
\arguments{
\item{pdf_file}{character vector of pdf files}

\item{verbose}{print warning messages? Default true}
}
\value{
tibble having
}
\description{
Parse many pdf files into nested section-page structure
}
