% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-section.R
\name{djt_parse_summary}
\alias{djt_parse_summary}
\title{Heuristics to DJT file into TOC table}
\usage{
djt_parse_summary(txt_summary)
}
\arguments{
\item{txt_summary}{subset of raw text that contains the summary}
}
\description{
Heuristics to DJT file into TOC table. Tested in TRT01, TRT02 and TRT17
}
