% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-section.R
\name{djt_remove_header}
\alias{djt_remove_header}
\title{remove header from DJT.}
\usage{
djt_remove_header(txt)
}
\arguments{
\item{txt}{DJT text.}
}
\value{
processed text
}
\description{
Tested in TRT01, TRT02 and TRT17
}
