% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-section.R
\name{djt_remove_regex}
\alias{djt_remove_regex}
\title{Remove regex from DJT}
\usage{
djt_remove_regex(txt, re)
}
\arguments{
\item{txt}{DJT text}

\item{re}{character atomic or regex object}
}
\value{
processed text
}
\description{
Remove sprcific regex from DJT
}
