% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-section.R
\name{djt_stack_columns}
\alias{djt_stack_columns}
\title{Stack columns}
\usage{
djt_stack_columns(txt, col1_init = 1, col1_end = 86, col2_init = 87,
  col2_end = -1)
}
\arguments{
\item{txt}{DJT text}

\item{col1_init}{index to start first column}

\item{col1_end}{index to end first column}

\item{col2_init}{index to start second column}

\item{col2_end}{index to end second column}
}
\value{
processed text
}
\description{
Stack two columns from each page text position
}
