% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-section.R
\name{djt_toc}
\alias{djt_toc}
\title{Gets toc from text of pdf file and text}
\usage{
djt_toc(txt, pdf_file)
}
\arguments{
\item{txt}{text after remove headers and footers and stack columns}

\item{pdf_file}{pdf file path to get the meta toc}
}
\value{
tibble containing sections
}
\description{
Gets summary of raw text, identifies sections and tries to
associate sections to each page. Also, compares with meta-toc
to check if there are any errors on the parsing process
}
