% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special-cleaners.R
\name{djt_trim}
\alias{djt_trim}
\title{Special cleaner
split raw text into escapes and trim output}
\usage{
djt_trim(txt_raw)
}
\description{
Special cleaner
split raw text into escapes and trim output
}
