% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djt.R
\name{download_djt}
\alias{download_djt}
\title{Download DJT PDFs based on date, booklet, and trt}
\usage{
download_djt(trt, path, booklet = "all", date_min = Sys.Date() - 1,
  date_max = date_min)
}
\arguments{
\item{trt}{Number between 1 and 24 indicating what TRT to download DJT from
or \code{"all"} for all TRTs}

\item{path}{Path where to save files}

\item{booklet}{One of \code{"jud"} (for judicial), \code{"adm"} (for administrative),
or \code{"all"} (for both)}

\item{date_min}{Lower bound for dates to be downloaded}

\item{date_max}{Upper bound for dates to be downloaded}
}
\value{
A \code{tibble} with identifiers and DJT URLs
}
\description{
Download DJT PDFs based on date, booklet, and trt
}
