% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-toc.R
\name{extract_part}
\alias{extract_part}
\title{extract name of the part from trimmed column based on djt_party_dict()}
\usage{
extract_part(column)
}
\description{
extract name of the part from trimmed column based on djt_party_dict()
}
