% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{find_all}
\alias{find_all}
\title{Find all occurrences of multiple patterns}
\usage{
find_all(string, ...)
}
\arguments{
\item{string}{Input string}

\item{...}{Patterns to be matched}
}
\description{
Find all occurrences of multiple patterns
}
