% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-meta-toc.R
\name{get_id_summary}
\alias{get_id_summary}
\title{a palavra sumário aparece várias vezes, por conta do "RITO SUMÁRIO"
mas ela aparece com 2 espaços antes uma vez só}
\usage{
get_id_summary(txt)
}
\description{
a palavra sumário aparece várias vezes, por conta do "RITO SUMÁRIO"
mas ela aparece com 2 espaços antes uma vez só
}
