% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{match_lawsuits}
\alias{match_lawsuits}
\title{Get lawsuits that match a pattern}
\usage{
match_lawsuits(string, pattern, only_words = TRUE, return_id = FALSE)
}
\arguments{
\item{string}{Input string}

\item{pattern}{Pattern to match}

\item{only_words}{Make sure pattern is surrounded by non-word characters
(see \code{\W} at \code{help("stringi-search-regex")})}

\item{return_id}{Whether to return only lawsuit IDs or the whole text of
the matched lawsuits}
}
\description{
Get lawsuits that match a pattern
}
