% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-meta-toc.R
\name{parse_single_section}
\alias{parse_single_section}
\title{Parse a section from TOC obtained from pdftools::pdf_toc()
This is a helper function for djt_meta_toc}
\usage{
parse_single_section(x)
}
\description{
Parse a section from TOC obtained from pdftools::pdf_toc()
This is a helper function for djt_meta_toc
}
