% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{pre_process}
\alias{pre_process}
\title{Preprocess text (remove headers and footers)}
\usage{
pre_process(string)
}
\arguments{
\item{string}{Input string}
}
\description{
Preprocess text (remove headers and footers)
}
