% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{remove_first_header}
\alias{remove_first_header}
\title{Remove text's first header}
\usage{
remove_first_header(string)
}
\arguments{
\item{string}{Input string}
}
\description{
Remove text's first header
}
