% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{remove_headers}
\alias{remove_headers}
\title{Remove text's headers (except first)}
\usage{
remove_headers(string)
}
\arguments{
\item{string}{Input string}
}
\description{
Remove text's headers (except first)
}
