% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{match_lawsuits}
\alias{match_lawsuits}
\title{Get lawsuits that match a pattern}
\usage{
match_lawsuits(string, pattern)
}
\arguments{
\item{string}{Input string}

\item{pattern}{Pattern to match}
}
\description{
Get lawsuits that match a pattern
}
